Execute auto scripts on new thread instead of blocking
