+++ src/xbot/common/autonomous/AutonomousScriptedCommand.java
@@ -2,26 +2,18 @@ package xbot.common.autonomous;
 
 import java.util.HashSet;
 import java.util.Set;
-import java.util.concurrent.ScheduledExecutorService;
 
 import org.apache.log4j.Logger;
-import org.mozilla.javascript.Context;
-import org.mozilla.javascript.Scriptable;
-import org.mozilla.javascript.ScriptableObject;
 
 import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.Scheduler;
-import xbot.common.command.BaseRobot;
 
 public class AutonomousScriptedCommand extends Command {
 
     static Logger log = Logger.getLogger(AutonomousScriptedCommand.class);
     
-    Context jsContext;
-    Scriptable jsScope;
-    
-    ScriptableObject robotInterfaceObject;
-    
+    private AutonomousScriptedCommandThread execThread;
+
     ScriptedCommandFactory availableCommandFactory;
     
     /**
@@ -39,50 +31,27 @@ public class AutonomousScriptedCommand extends Command {
     }
     
     private void initializeScriptEngine() {
-        // TODO: Figure out if this will work with multiple simultaneous scripted commands
-        jsContext = Context.enter();
-        jsScope = jsContext.initStandardObjects();
-        robotInterfaceObject = new ScriptableObject() {
-            
-            @Override
-            public String getClassName() {
-                return "RobotInterface";
-            }
-        };
-        jsScope.put("robot",  jsScope, robotInterfaceObject);
-        robotInterfaceObject.put("requireCommands", robotInterfaceObject,
-                new RequireCommandsFunction(commandNames -> this.requireCommands(commandNames)));
-        
+        execThread = new AutonomousScriptedCommandThread(this, availableCommandFactory);
+        execThread.initializeScriptEngine();
     }
     
+    /**
+     * Executes the given script string.
+     * WARNING: This should only be used for debugging. Real robots should use
+     * the constructor to pass in a target file.
+     * 
+     * @param scriptText
+     * @param scriptName
+     */
     public void executeScriptFromString(String scriptText, String scriptName) {
-        this.jsContext.evaluateString(jsScope, scriptText, scriptName, 1, null);
+        execThread.executeScriptFromString(scriptText, scriptName);
     }
     
-    private void requireCommands(String[] commandNames) {
-        if(this.jsContext == null) {
-            log.error("An attempt was made to require a command after the JS context had been"
-                    + "de-initialized. Ignoring this attempt.");
-            return;
-        }
-        
-        for(String commandTypeName : commandNames) {
-            log.info("Autonomous script required command " + commandTypeName);
-            
-            ScriptedCommandProvider commandProvider = this.availableCommandFactory.getProviderForName(commandTypeName);
-            
-            if(commandProvider == null) {
-                log.error("Unable to get provider for command type " + commandTypeName + "!");
-                continue;
-            }
-            
-            InvokeCommandFunction commandInvoker = new InvokeCommandFunction(commandProvider, command -> this.invokeCommand(command));
-            
-            jsScope.put("invoke" + commandTypeName, robotInterfaceObject, commandInvoker);
-        }
-    }
-    
-    private void invokeCommand(Command command) {
+    /**
+     * For internal use only!
+     * @param command
+     */
+    public synchronized void invokeCommand(Command command) {
         Scheduler.getInstance().add(command);
         if(this.invokedCommands == null)
             this.invokedCommands = new HashSet<Command>();
@@ -92,9 +61,7 @@ public class AutonomousScriptedCommand extends Command {
     
     @Override
     protected void initialize() {
-        // If we have already run and been interrupted once, we need to re-init
-        //  the context.
-        if(jsContext == null)
+        if(execThread == null)
             initializeScriptEngine();
     }
 
@@ -112,6 +79,7 @@ public class AutonomousScriptedCommand extends Command {
 
     @Override
     protected void end() {
+        this.execThread.interrupt();
         if(this.invokedCommands != null) {
             for(Command command : invokedCommands) {
                 command.cancel();
@@ -119,10 +87,7 @@ public class AutonomousScriptedCommand extends Command {
             
             this.invokedCommands = null;
         }
-        
-        Context.exit();
-        this.jsContext = null;
-        this.jsScope = null;
+        this.execThread = null;
     }
 
     @Override
